[Java][Spring] constructor with required args
