+++ samples/server/petstore/springboot/.openapi-generator-ignore
@@ -22,4 +22,4 @@
 # Then explicitly reverse the ignore rule for a single file:
 #!docs/README.md
 
-src/test/java/org/openapitools/OpenApiGeneratorApplicationTests.java
+src/test/java/org/openapitools/JacksonTest.java
