+++ samples/server/petstore/springboot/.openapi-generator/FILES
@@ -69,3 +69,4 @@ src/main/java/org/openapitools/model/XmlItemDto.java
 src/main/resources/application.properties
 src/main/resources/openapi.yaml
 src/main/resources/static/swagger-ui.html
+src/test/java/org/openapitools/OpenApiGeneratorApplicationTests.java
