--- samples/server/petstore/springboot/src/main/java/org/openapitools/model/BigCatAllOf.java
@@ -1,126 +1,0 @@
-package org.openapitools.model;
-
-import java.net.URI;
-import java.util.Objects;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonTypeName;
-import com.fasterxml.jackson.annotation.JsonValue;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import org.openapitools.jackson.nullable.JsonNullable;
-import java.time.OffsetDateTime;
-import javax.validation.Valid;
-import javax.validation.constraints.*;
-
-
-import java.util.*;
-import javax.annotation.Generated;
-
-/**
- * BigCatAllOf
- */
-
-@JsonTypeName("BigCat_allOf")
-@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
-public class BigCatAllOf {
-
-  /**
-   * Gets or Sets kind
-   */
-  public enum KindEnum {
-    LIONS("lions"),
-    
-    TIGERS("tigers"),
-    
-    LEOPARDS("leopards"),
-    
-    JAGUARS("jaguars");
-
-    private String value;
-
-    KindEnum(String value) {
-      this.value = value;
-    }
-
-    @JsonValue
-    public String getValue() {
-      return value;
-    }
-
-    @Override
-    public String toString() {
-      return String.valueOf(value);
-    }
-
-    @JsonCreator
-    public static KindEnum fromValue(String value) {
-      for (KindEnum b : KindEnum.values()) {
-        if (b.value.equals(value)) {
-          return b;
-        }
-      }
-      throw new IllegalArgumentException("Unexpected value '" + value + "'");
-    }
-  }
-
-  @JsonProperty("kind")
-  private KindEnum kind;
-
-  public BigCatAllOf kind(KindEnum kind) {
-    this.kind = kind;
-    return this;
-  }
-
-  /**
-   * Get kind
-   * @return kind
-  */
-  
-  @ApiModelProperty(value = "")
-  public KindEnum getKind() {
-    return kind;
-  }
-
-  public void setKind(KindEnum kind) {
-    this.kind = kind;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    BigCatAllOf bigCatAllOf = (BigCatAllOf) o;
-    return Objects.equals(this.kind, bigCatAllOf.kind);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(kind);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class BigCatAllOf {\n");
-    sb.append("    kind: ").append(toIndentedString(kind)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
-}
-
