--- samples/server/petstore/springboot/src/main/java/org/openapitools/model/Cat.java
@@ -1,110 +1,0 @@
-package org.openapitools.model;
-
-import java.net.URI;
-import java.util.Objects;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import org.openapitools.model.Animal;
-import org.openapitools.model.BigCat;
-import org.openapitools.jackson.nullable.JsonNullable;
-import java.time.OffsetDateTime;
-import javax.validation.Valid;
-import javax.validation.constraints.*;
-
-
-import java.util.*;
-import javax.annotation.Generated;
-
-/**
- * Cat
- */
-
-@JsonIgnoreProperties(
-  value = "className", // ignore manually set className, it will be automatically generated by Jackson during serialization
-  allowSetters = true // allows the className to be set during deserialization
-)
-@JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "className", visible = true)
-@JsonSubTypes({
-  @JsonSubTypes.Type(value = BigCat.class, name = "BigCat")
-})
-
-@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
-public class Cat extends Animal {
-
-  @JsonProperty("declawed")
-  private Boolean declawed;
-
-  public Cat declawed(Boolean declawed) {
-    this.declawed = declawed;
-    return this;
-  }
-
-  /**
-   * Get declawed
-   * @return declawed
-  */
-  
-  @ApiModelProperty(value = "")
-  public Boolean getDeclawed() {
-    return declawed;
-  }
-
-  public void setDeclawed(Boolean declawed) {
-    this.declawed = declawed;
-  }
-
-  public Cat className(String className) {
-    super.setClassName(className);
-    return this;
-  }
-
-  public Cat color(String color) {
-    super.setColor(color);
-    return this;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    Cat cat = (Cat) o;
-    return Objects.equals(this.declawed, cat.declawed) &&
-        super.equals(o);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(declawed, super.hashCode());
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class Cat {\n");
-    sb.append("    ").append(toIndentedString(super.toString())).append("\n");
-    sb.append("    declawed: ").append(toIndentedString(declawed)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
-}
-
