--- samples/server/petstore/springboot/src/main/java/org/openapitools/model/CatAllOf.java
@@ -1,86 +1,0 @@
-package org.openapitools.model;
-
-import java.net.URI;
-import java.util.Objects;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonTypeName;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import org.openapitools.jackson.nullable.JsonNullable;
-import java.time.OffsetDateTime;
-import javax.validation.Valid;
-import javax.validation.constraints.*;
-
-
-import java.util.*;
-import javax.annotation.Generated;
-
-/**
- * CatAllOf
- */
-
-@JsonTypeName("Cat_allOf")
-@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
-public class CatAllOf {
-
-  @JsonProperty("declawed")
-  private Boolean declawed;
-
-  public CatAllOf declawed(Boolean declawed) {
-    this.declawed = declawed;
-    return this;
-  }
-
-  /**
-   * Get declawed
-   * @return declawed
-  */
-  
-  @ApiModelProperty(value = "")
-  public Boolean getDeclawed() {
-    return declawed;
-  }
-
-  public void setDeclawed(Boolean declawed) {
-    this.declawed = declawed;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    CatAllOf catAllOf = (CatAllOf) o;
-    return Objects.equals(this.declawed, catAllOf.declawed);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(declawed);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class CatAllOf {\n");
-    sb.append("    declawed: ").append(toIndentedString(declawed)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
-}
-
