--- samples/server/petstore/springboot/src/main/java/org/openapitools/model/ClassModel.java
@@ -1,85 +1,0 @@
-package org.openapitools.model;
-
-import java.net.URI;
-import java.util.Objects;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import org.openapitools.jackson.nullable.JsonNullable;
-import java.time.OffsetDateTime;
-import javax.validation.Valid;
-import javax.validation.constraints.*;
-
-
-import java.util.*;
-import javax.annotation.Generated;
-
-/**
- * Model for testing model with \&quot;_class\&quot; property
- */
-
-@ApiModel(description = "Model for testing model with \"_class\" property")
-@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
-public class ClassModel {
-
-  @JsonProperty("_class")
-  private String propertyClass;
-
-  public ClassModel propertyClass(String propertyClass) {
-    this.propertyClass = propertyClass;
-    return this;
-  }
-
-  /**
-   * Get propertyClass
-   * @return propertyClass
-  */
-  
-  @ApiModelProperty(value = "")
-  public String getPropertyClass() {
-    return propertyClass;
-  }
-
-  public void setPropertyClass(String propertyClass) {
-    this.propertyClass = propertyClass;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    ClassModel classModel = (ClassModel) o;
-    return Objects.equals(this.propertyClass, classModel.propertyClass);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(propertyClass);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class ClassModel {\n");
-    sb.append("    propertyClass: ").append(toIndentedString(propertyClass)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
-}
-
