--- samples/server/petstore/springboot/src/main/java/org/openapitools/model/Dog.java
@@ -1,101 +1,0 @@
-package org.openapitools.model;
-
-import java.net.URI;
-import java.util.Objects;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import org.openapitools.model.Animal;
-import org.openapitools.jackson.nullable.JsonNullable;
-import java.time.OffsetDateTime;
-import javax.validation.Valid;
-import javax.validation.constraints.*;
-
-
-import java.util.*;
-import javax.annotation.Generated;
-
-/**
- * Dog
- */
-
-
-@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
-public class Dog extends Animal {
-
-  @JsonProperty("breed")
-  private String breed;
-
-  public Dog breed(String breed) {
-    this.breed = breed;
-    return this;
-  }
-
-  /**
-   * Get breed
-   * @return breed
-  */
-  
-  @ApiModelProperty(value = "")
-  public String getBreed() {
-    return breed;
-  }
-
-  public void setBreed(String breed) {
-    this.breed = breed;
-  }
-
-  public Dog className(String className) {
-    super.setClassName(className);
-    return this;
-  }
-
-  public Dog color(String color) {
-    super.setColor(color);
-    return this;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    Dog dog = (Dog) o;
-    return Objects.equals(this.breed, dog.breed) &&
-        super.equals(o);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(breed, super.hashCode());
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class Dog {\n");
-    sb.append("    ").append(toIndentedString(super.toString())).append("\n");
-    sb.append("    breed: ").append(toIndentedString(breed)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
-}
-
