--- samples/server/petstore/springboot/src/main/java/org/openapitools/model/EnumArrays.java
@@ -1,190 +1,0 @@
-package org.openapitools.model;
-
-import java.net.URI;
-import java.util.Objects;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonValue;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import java.util.ArrayList;
-import java.util.List;
-import org.openapitools.jackson.nullable.JsonNullable;
-import java.time.OffsetDateTime;
-import javax.validation.Valid;
-import javax.validation.constraints.*;
-
-
-import java.util.*;
-import javax.annotation.Generated;
-
-/**
- * EnumArrays
- */
-
-@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
-public class EnumArrays {
-
-  /**
-   * Gets or Sets justSymbol
-   */
-  public enum JustSymbolEnum {
-    GREATER_THAN_OR_EQUAL_TO(">="),
-    
-    DOLLAR("$");
-
-    private String value;
-
-    JustSymbolEnum(String value) {
-      this.value = value;
-    }
-
-    @JsonValue
-    public String getValue() {
-      return value;
-    }
-
-    @Override
-    public String toString() {
-      return String.valueOf(value);
-    }
-
-    @JsonCreator
-    public static JustSymbolEnum fromValue(String value) {
-      for (JustSymbolEnum b : JustSymbolEnum.values()) {
-        if (b.value.equals(value)) {
-          return b;
-        }
-      }
-      throw new IllegalArgumentException("Unexpected value '" + value + "'");
-    }
-  }
-
-  @JsonProperty("just_symbol")
-  private JustSymbolEnum justSymbol;
-
-  /**
-   * Gets or Sets arrayEnum
-   */
-  public enum ArrayEnumEnum {
-    FISH("fish"),
-    
-    CRAB("crab");
-
-    private String value;
-
-    ArrayEnumEnum(String value) {
-      this.value = value;
-    }
-
-    @JsonValue
-    public String getValue() {
-      return value;
-    }
-
-    @Override
-    public String toString() {
-      return String.valueOf(value);
-    }
-
-    @JsonCreator
-    public static ArrayEnumEnum fromValue(String value) {
-      for (ArrayEnumEnum b : ArrayEnumEnum.values()) {
-        if (b.value.equals(value)) {
-          return b;
-        }
-      }
-      throw new IllegalArgumentException("Unexpected value '" + value + "'");
-    }
-  }
-
-  @JsonProperty("array_enum")
-  @Valid
-  private List<ArrayEnumEnum> arrayEnum = null;
-
-  public EnumArrays justSymbol(JustSymbolEnum justSymbol) {
-    this.justSymbol = justSymbol;
-    return this;
-  }
-
-  /**
-   * Get justSymbol
-   * @return justSymbol
-  */
-  
-  @ApiModelProperty(value = "")
-  public JustSymbolEnum getJustSymbol() {
-    return justSymbol;
-  }
-
-  public void setJustSymbol(JustSymbolEnum justSymbol) {
-    this.justSymbol = justSymbol;
-  }
-
-  public EnumArrays arrayEnum(List<ArrayEnumEnum> arrayEnum) {
-    this.arrayEnum = arrayEnum;
-    return this;
-  }
-
-  public EnumArrays addArrayEnumItem(ArrayEnumEnum arrayEnumItem) {
-    if (this.arrayEnum == null) {
-      this.arrayEnum = new ArrayList<>();
-    }
-    this.arrayEnum.add(arrayEnumItem);
-    return this;
-  }
-
-  /**
-   * Get arrayEnum
-   * @return arrayEnum
-  */
-  
-  @ApiModelProperty(value = "")
-  public List<ArrayEnumEnum> getArrayEnum() {
-    return arrayEnum;
-  }
-
-  public void setArrayEnum(List<ArrayEnumEnum> arrayEnum) {
-    this.arrayEnum = arrayEnum;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    EnumArrays enumArrays = (EnumArrays) o;
-    return Objects.equals(this.justSymbol, enumArrays.justSymbol) &&
-        Objects.equals(this.arrayEnum, enumArrays.arrayEnum);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(justSymbol, arrayEnum);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class EnumArrays {\n");
-    sb.append("    justSymbol: ").append(toIndentedString(justSymbol)).append("\n");
-    sb.append("    arrayEnum: ").append(toIndentedString(arrayEnum)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
-}
-
