--- samples/server/petstore/springboot/src/main/java/org/openapitools/model/MapTest.java
@@ -1,230 +1,0 @@
-package org.openapitools.model;
-
-import java.net.URI;
-import java.util.Objects;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonValue;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import java.util.HashMap;
-import java.util.Map;
-import org.openapitools.jackson.nullable.JsonNullable;
-import java.time.OffsetDateTime;
-import javax.validation.Valid;
-import javax.validation.constraints.*;
-
-
-import java.util.*;
-import javax.annotation.Generated;
-
-/**
- * MapTest
- */
-
-@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
-public class MapTest {
-
-  @JsonProperty("map_map_of_string")
-  @Valid
-  private Map<String, Map<String, String>> mapMapOfString = null;
-
-  /**
-   * Gets or Sets inner
-   */
-  public enum InnerEnum {
-    UPPER("UPPER"),
-    
-    LOWER("lower");
-
-    private String value;
-
-    InnerEnum(String value) {
-      this.value = value;
-    }
-
-    @JsonValue
-    public String getValue() {
-      return value;
-    }
-
-    @Override
-    public String toString() {
-      return String.valueOf(value);
-    }
-
-    @JsonCreator
-    public static InnerEnum fromValue(String value) {
-      for (InnerEnum b : InnerEnum.values()) {
-        if (b.value.equals(value)) {
-          return b;
-        }
-      }
-      throw new IllegalArgumentException("Unexpected value '" + value + "'");
-    }
-  }
-
-  @JsonProperty("map_of_enum_string")
-  @Valid
-  private Map<String, InnerEnum> mapOfEnumString = null;
-
-  @JsonProperty("direct_map")
-  @Valid
-  private Map<String, Boolean> directMap = null;
-
-  @JsonProperty("indirect_map")
-  @Valid
-  private Map<String, Boolean> indirectMap = null;
-
-  public MapTest mapMapOfString(Map<String, Map<String, String>> mapMapOfString) {
-    this.mapMapOfString = mapMapOfString;
-    return this;
-  }
-
-  public MapTest putMapMapOfStringItem(String key, Map<String, String> mapMapOfStringItem) {
-    if (this.mapMapOfString == null) {
-      this.mapMapOfString = new HashMap<>();
-    }
-    this.mapMapOfString.put(key, mapMapOfStringItem);
-    return this;
-  }
-
-  /**
-   * Get mapMapOfString
-   * @return mapMapOfString
-  */
-  @Valid 
-  @ApiModelProperty(value = "")
-  public Map<String, Map<String, String>> getMapMapOfString() {
-    return mapMapOfString;
-  }
-
-  public void setMapMapOfString(Map<String, Map<String, String>> mapMapOfString) {
-    this.mapMapOfString = mapMapOfString;
-  }
-
-  public MapTest mapOfEnumString(Map<String, InnerEnum> mapOfEnumString) {
-    this.mapOfEnumString = mapOfEnumString;
-    return this;
-  }
-
-  public MapTest putMapOfEnumStringItem(String key, InnerEnum mapOfEnumStringItem) {
-    if (this.mapOfEnumString == null) {
-      this.mapOfEnumString = new HashMap<>();
-    }
-    this.mapOfEnumString.put(key, mapOfEnumStringItem);
-    return this;
-  }
-
-  /**
-   * Get mapOfEnumString
-   * @return mapOfEnumString
-  */
-  
-  @ApiModelProperty(value = "")
-  public Map<String, InnerEnum> getMapOfEnumString() {
-    return mapOfEnumString;
-  }
-
-  public void setMapOfEnumString(Map<String, InnerEnum> mapOfEnumString) {
-    this.mapOfEnumString = mapOfEnumString;
-  }
-
-  public MapTest directMap(Map<String, Boolean> directMap) {
-    this.directMap = directMap;
-    return this;
-  }
-
-  public MapTest putDirectMapItem(String key, Boolean directMapItem) {
-    if (this.directMap == null) {
-      this.directMap = new HashMap<>();
-    }
-    this.directMap.put(key, directMapItem);
-    return this;
-  }
-
-  /**
-   * Get directMap
-   * @return directMap
-  */
-  
-  @ApiModelProperty(value = "")
-  public Map<String, Boolean> getDirectMap() {
-    return directMap;
-  }
-
-  public void setDirectMap(Map<String, Boolean> directMap) {
-    this.directMap = directMap;
-  }
-
-  public MapTest indirectMap(Map<String, Boolean> indirectMap) {
-    this.indirectMap = indirectMap;
-    return this;
-  }
-
-  public MapTest putIndirectMapItem(String key, Boolean indirectMapItem) {
-    if (this.indirectMap == null) {
-      this.indirectMap = new HashMap<>();
-    }
-    this.indirectMap.put(key, indirectMapItem);
-    return this;
-  }
-
-  /**
-   * Get indirectMap
-   * @return indirectMap
-  */
-  
-  @ApiModelProperty(value = "")
-  public Map<String, Boolean> getIndirectMap() {
-    return indirectMap;
-  }
-
-  public void setIndirectMap(Map<String, Boolean> indirectMap) {
-    this.indirectMap = indirectMap;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    MapTest mapTest = (MapTest) o;
-    return Objects.equals(this.mapMapOfString, mapTest.mapMapOfString) &&
-        Objects.equals(this.mapOfEnumString, mapTest.mapOfEnumString) &&
-        Objects.equals(this.directMap, mapTest.directMap) &&
-        Objects.equals(this.indirectMap, mapTest.indirectMap);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(mapMapOfString, mapOfEnumString, directMap, indirectMap);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class MapTest {\n");
-    sb.append("    mapMapOfString: ").append(toIndentedString(mapMapOfString)).append("\n");
-    sb.append("    mapOfEnumString: ").append(toIndentedString(mapOfEnumString)).append("\n");
-    sb.append("    directMap: ").append(toIndentedString(directMap)).append("\n");
-    sb.append("    indirectMap: ").append(toIndentedString(indirectMap)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
-}
-
