--- samples/server/petstore/springboot/src/main/java/org/openapitools/model/Model200Response.java
@@ -1,111 +1,0 @@
-package org.openapitools.model;
-
-import java.net.URI;
-import java.util.Objects;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonTypeName;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import org.openapitools.jackson.nullable.JsonNullable;
-import java.time.OffsetDateTime;
-import javax.validation.Valid;
-import javax.validation.constraints.*;
-
-
-import java.util.*;
-import javax.annotation.Generated;
-
-/**
- * Model for testing model name starting with number
- */
-
-@ApiModel(description = "Model for testing model name starting with number")
-@JsonTypeName("200_response")
-@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
-public class Model200Response {
-
-  @JsonProperty("name")
-  private Integer name;
-
-  @JsonProperty("class")
-  private String propertyClass;
-
-  public Model200Response name(Integer name) {
-    this.name = name;
-    return this;
-  }
-
-  /**
-   * Get name
-   * @return name
-  */
-  
-  @ApiModelProperty(value = "")
-  public Integer getName() {
-    return name;
-  }
-
-  public void setName(Integer name) {
-    this.name = name;
-  }
-
-  public Model200Response propertyClass(String propertyClass) {
-    this.propertyClass = propertyClass;
-    return this;
-  }
-
-  /**
-   * Get propertyClass
-   * @return propertyClass
-  */
-  
-  @ApiModelProperty(value = "")
-  public String getPropertyClass() {
-    return propertyClass;
-  }
-
-  public void setPropertyClass(String propertyClass) {
-    this.propertyClass = propertyClass;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    Model200Response _200Response = (Model200Response) o;
-    return Objects.equals(this.name, _200Response.name) &&
-        Objects.equals(this.propertyClass, _200Response.propertyClass);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(name, propertyClass);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class Model200Response {\n");
-    sb.append("    name: ").append(toIndentedString(name)).append("\n");
-    sb.append("    propertyClass: ").append(toIndentedString(propertyClass)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
-}
-
