--- samples/server/petstore/springboot/src/main/java/org/openapitools/model/ModelList.java
@@ -1,86 +1,0 @@
-package org.openapitools.model;
-
-import java.net.URI;
-import java.util.Objects;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonTypeName;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import org.openapitools.jackson.nullable.JsonNullable;
-import java.time.OffsetDateTime;
-import javax.validation.Valid;
-import javax.validation.constraints.*;
-
-
-import java.util.*;
-import javax.annotation.Generated;
-
-/**
- * ModelList
- */
-
-@JsonTypeName("List")
-@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
-public class ModelList {
-
-  @JsonProperty("123-list")
-  private String _123List;
-
-  public ModelList _123List(String _123List) {
-    this._123List = _123List;
-    return this;
-  }
-
-  /**
-   * Get _123List
-   * @return _123List
-  */
-  
-  @ApiModelProperty(value = "")
-  public String get123List() {
-    return _123List;
-  }
-
-  public void set123List(String _123List) {
-    this._123List = _123List;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    ModelList _list = (ModelList) o;
-    return Objects.equals(this._123List, _list._123List);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(_123List);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class ModelList {\n");
-    sb.append("    _123List: ").append(toIndentedString(_123List)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
-}
-
