--- samples/server/petstore/springboot/src/main/java/org/openapitools/model/ModelReturn.java
@@ -1,87 +1,0 @@
-package org.openapitools.model;
-
-import java.net.URI;
-import java.util.Objects;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonTypeName;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import org.openapitools.jackson.nullable.JsonNullable;
-import java.time.OffsetDateTime;
-import javax.validation.Valid;
-import javax.validation.constraints.*;
-
-
-import java.util.*;
-import javax.annotation.Generated;
-
-/**
- * Model for testing reserved words
- */
-
-@ApiModel(description = "Model for testing reserved words")
-@JsonTypeName("Return")
-@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
-public class ModelReturn {
-
-  @JsonProperty("return")
-  private Integer _return;
-
-  public ModelReturn _return(Integer _return) {
-    this._return = _return;
-    return this;
-  }
-
-  /**
-   * Get _return
-   * @return _return
-  */
-  
-  @ApiModelProperty(value = "")
-  public Integer getReturn() {
-    return _return;
-  }
-
-  public void setReturn(Integer _return) {
-    this._return = _return;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    ModelReturn _return = (ModelReturn) o;
-    return Objects.equals(this._return, _return._return);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(_return);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class ModelReturn {\n");
-    sb.append("    _return: ").append(toIndentedString(_return)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
-}
-
