--- samples/server/petstore/springboot/src/main/java/org/openapitools/model/OuterEnum.java
@@ -1,57 +1,0 @@
-package org.openapitools.model;
-
-import java.net.URI;
-import java.util.Objects;
-import com.fasterxml.jackson.annotation.JsonValue;
-import org.openapitools.jackson.nullable.JsonNullable;
-import java.time.OffsetDateTime;
-import javax.validation.Valid;
-import javax.validation.constraints.*;
-
-
-import java.util.*;
-import javax.annotation.Generated;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonValue;
-
-/**
- * Gets or Sets OuterEnum
- */
-
-@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
-public enum OuterEnum {
-  
-  PLACED("placed"),
-  
-  APPROVED("approved"),
-  
-  DELIVERED("delivered");
-
-  private String value;
-
-  OuterEnum(String value) {
-    this.value = value;
-  }
-
-  @JsonValue
-  public String getValue() {
-    return value;
-  }
-
-  @Override
-  public String toString() {
-    return String.valueOf(value);
-  }
-
-  @JsonCreator
-  public static OuterEnum fromValue(String value) {
-    for (OuterEnum b : OuterEnum.values()) {
-      if (b.value.equals(value)) {
-        return b;
-      }
-    }
-    throw new IllegalArgumentException("Unexpected value '" + value + "'");
-  }
-}
-
