--- samples/server/petstore/springboot/src/main/java/org/openapitools/model/ReadOnlyFirst.java
@@ -1,108 +1,0 @@
-package org.openapitools.model;
-
-import java.net.URI;
-import java.util.Objects;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import org.openapitools.jackson.nullable.JsonNullable;
-import java.time.OffsetDateTime;
-import javax.validation.Valid;
-import javax.validation.constraints.*;
-
-
-import java.util.*;
-import javax.annotation.Generated;
-
-/**
- * ReadOnlyFirst
- */
-
-@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
-public class ReadOnlyFirst {
-
-  @JsonProperty("bar")
-  private String bar;
-
-  @JsonProperty("baz")
-  private String baz;
-
-  public ReadOnlyFirst bar(String bar) {
-    this.bar = bar;
-    return this;
-  }
-
-  /**
-   * Get bar
-   * @return bar
-  */
-  
-  @ApiModelProperty(readOnly = true, value = "")
-  public String getBar() {
-    return bar;
-  }
-
-  public void setBar(String bar) {
-    this.bar = bar;
-  }
-
-  public ReadOnlyFirst baz(String baz) {
-    this.baz = baz;
-    return this;
-  }
-
-  /**
-   * Get baz
-   * @return baz
-  */
-  
-  @ApiModelProperty(value = "")
-  public String getBaz() {
-    return baz;
-  }
-
-  public void setBaz(String baz) {
-    this.baz = baz;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    ReadOnlyFirst readOnlyFirst = (ReadOnlyFirst) o;
-    return Objects.equals(this.bar, readOnlyFirst.bar) &&
-        Objects.equals(this.baz, readOnlyFirst.baz);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(bar, baz);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class ReadOnlyFirst {\n");
-    sb.append("    bar: ").append(toIndentedString(bar)).append("\n");
-    sb.append("    baz: ").append(toIndentedString(baz)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
-}
-
